requirements.txt
Flask==2.3.3
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
flask-jwt-extended==4.6.0
